pyproject.toml
[tool.black]
line-length = 120

[tool.isort]
profile = "black"
line_length = 120
skip_gitignore = true

[tool.poetry]
name = "academyruins-api"
version = "0.5.0"
description = ""
authors = ["Václav Luňák < >"]
readme = "README.md"

[tool.poetry.scripts]
academyruins = "src.cli_scripts.run:app"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = { extras = ["all"], version = "^0.103.2" }
APScheduler = "^3.9.1"
boto3 = "^1.24.84"
pydantic = "^2.4.2"
requests = "^2.28.1"
SQLAlchemy = "^1.4.41"
psycopg2-binary = "^2.9.3"
hjson = "^3.1.0"
asyncio = "^3.4.3"
beautifulsoup4 = "^4.11.1"
thefuzz = {extras = ["speedup"], version = "^0.19.0"}
tika = "^1.24"
alembic = "^1.8.1"


[tool.poetry.group.dev.dependencies]
python-dotenv = "^0.21.0"
black = "^22.8.0"
alembic = "^1.8.1"
flake8 = "^6.0.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
